requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
pytest>=7.4.0
pytest-django>=4.5.0
django-allauth>=0.57.0
psycopg2-binary>=2.9.0
pytz>=2024.1
